This PR simplifies the return result of step_inner function.

Using `Result<Never, InterpreterResult>` as return result will cause 
unnecessary overhead (mostly additional memory operations). Since 
step_inner is a hot function, I propose to simplify the return type of 
step_inner function and handle the error results manually.

Some benchmark results I have:

Without Optimization (Original Version):
test rng_u256          ... bench:  68,395,922 ns/iter (+/- 19,520,348)
test rng_usize         ... bench:  46,822,703 ns/iter (+/- 19,869,026)
test simple_loop_u256  ... bench:  18,807,019 ns/iter (+/- 5,683,907)
test simple_loop_usize ... bench:  13,812,568 ns/iter (+/- 3,310,862)

With Optimization:
test rng_u256          ... bench:  53,654,255 ns/iter (+/- 12,985,727)
test rng_usize         ... bench:  36,481,886 ns/iter (+/- 13,044,046)
test simple_loop_u256  ... bench:  15,010,101 ns/iter (+/- 5,306,132)
test simple_loop_usize ... bench:  10,584,103 ns/iter (+/- 2,439,375)

Some perf results I have:
 
<img width="1190" alt="image" src="https://user-images.githubusercontent.com/5557706/62407383-4a989880-b5ea-11e9-81a9-b4f800f6b0df.png">