@dvdplm 2019-08-05
Do you mind including the benchmark code in the PR?

Removing the error conversion and all uses of `?` makes the code less readable. Can you elaborate on why the return type change necessitates changing the error handling like this?