@Leeleo3x 2019-08-05
The benchmark I'm using is included in current Parity code base. To run benchmarks:
```
cargo bench --package evmbin
```

I do understand `?` operator makes the code more readable. However, the `?` operator calls `into_result` function and wraps the result into `Result` type which causes additional memory operations. This is also because the size of `InterpreterResult` is large and need to be stored in memory to move around.