@ordian 2019-08-05
> Do you mind including the benchmark code in the PR?

I think the benches are from `evmbin/benches/mod.rs`.

> Removing the error conversion and all uses of ? makes the code less readable. Can you elaborate on why the return type change necessitates changing the error handling like this?

That's a known issue https://github.com/rust-lang/rust/issues/37939