@dvdplm 2019-08-05
> That's a known issue [rust-lang/rust#37939](https://github.com/rust-lang/rust/issues/37939)

I had no idea, thank you for educating me.